services/auth_service/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.8.2
sqlmodel==0.0.22
sqlite-utils==3.37

